creacion del controlador de usuario y sus rutas
