--- src/middlewares/usuario-validator.js
@@ -1,67 +1,0 @@
-import { body, param } from "express-validator";
-import { emailExists, userExists } from "../helpers/db-validators.js";
-import { validarCampos } from "./validate-campos.js";
-import { deleteFileOnError } from "./delete-file-on-error.js";
-import { handleErrors } from "./haddle-error.js";
-
-export const registerValidator = [
-    body("nombre").notEmpty().withMessage("El nombre es requerido"),
-    body("email").notEmpty().withMessage("El email es requerido"),
-    body("email").isEmail().withMessage("No es un email válido"),
-    body("email").custom(emailExists),
-    body("password").isStrongPassword({
-        minLength: 8,
-        minLowercase: 1,
-        minUppercase: 1,
-        minNumbers: 1,
-        minSymbols: 1
-    }).withMessage("El password debe contener al menos 8 caracteres, una letra minúscula, una letra mayúscula, un número y un símbolo"),
-    validarCampos,
-    deleteFileOnError,
-    handleErrors
-];
-
-export const loginValidator = [
-    body("email").notEmpty().withMessage("El email es requerido"),
-    body("email").isEmail().withMessage("No es un email válido"),
-    body("password").isLength({ min: 8 }).withMessage("El password debe contener al menos 8 caracteres"),
-    validarCampos,
-    handleErrors
-];
-
-export const getUserByIdValidator = [
-    param("uid").isMongoId().withMessage("No es un ID válido de MongoDB"),
-    param("uid").custom(userExists),
-    validarCampos,
-    handleErrors
-];
-
-export const deleteUserValidator = [
-    param("uid").isMongoId().withMessage("No es un ID válido de MongoDB"),
-    param("uid").custom(userExists),
-    validarCampos,
-    handleErrors
-];
-
-export const updatePasswordValidator = [
-    param("uid").isMongoId().withMessage("No es un ID válido de MongoDB"),
-    param("uid").custom(userExists),
-    body("newPassword").isLength({ min: 8 }).withMessage("El password debe contener al menos 8 caracteres"),
-    validarCampos,
-    handleErrors
-];
-
-export const updateUserValidator = [
-    param("uid", "No es un ID válido").isMongoId(),
-    param("uid").custom(userExists),
-    validarCampos,
-    handleErrors
-];
-
-export const updateProfilePictureValidator = [
-    param("uid").isMongoId().withMessage("No es un ID válido de MongoDB"),
-    param("uid").custom(userExists),
-    validarCampos,
-    deleteFileOnError,
-    handleErrors
-];
